requirements.txt
pyautogui
playwright
vosk
flite
ollama
pytest-playwright
pytesseract
pillow
numpy
fastapi
uvicorn
click
asyncio
pydantic
python-multipart